Test loading DBDefs before starting server
It happened that a server instance would continuously error because it
did not start with a valid `lib/DBDefs.pm`.  This file is deployed in
production using consul-template which may not provide the correct
configuration occasionally.

This patch just tries loading DBDefs module before running the server.
If the module cannot be loaded, consul-template will try starting a new
instance automatically, hopefully using a correct configuration provided
by consul-agent in the meantime.

Disclaimer: This does NOT test anything beyond loading DBDefs module such as:
validity of configuration values, connection to the database or any other
service.

Impl. note: Server won’t reload `lib/DBDefs.pm` if it changed, so there is no
need to check DBDefs all the time while running. Consul is in charge of
stopping/restarting server.